# src/docs/swagger.yaml
definitions:
  forms.Result:
    properties:
      error:
        example: some error note
        type: string
      result:
        example: some success note
        type: string
      status:
        example: OK
        type: string
    type: object
  forms.User:
    properties:
      email:
        example: hello@example.com
        type: string
      id:
        example: 5125112
        type: integer
      password:
        example: 1fsgh2rfafas
        type: string
      username:
        example: Martin
        type: string
    type: object
...
  models.Track:
    properties:
      albumId:
        example: 143
        type: integer
      authorId:
        example: 121
        type: integer
      countLikes:
        example: 54
        type: integer
      countListening:
        example: 15632
        type: integer
      coverId:
        example: 254
        type: integer
      duration:
        example: 180
        type: integer
      id:
        example: 777
        type: integer
      mp4:
        example: assets/track_1.mp4
        type: string
      title:
        example: Rain
        type: string
    type: object
  utils.Error:
    properties:
      err:
        type: string
    type: object
  utils.Success:
    properties:
      result: {}
    type: object
info:
  contact: {}
paths:
...
  /net/v1/track/:
    get:
      consumes:
        - application/json
      description: getting all tracks
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetTracks
      tags:
        - track
    post:
      consumes:
      - application/json
      description: creating new track
      parameters:
        - description: params of new track. Id will be set automatically.
          in: body
          name: Track
          required: true
          schema:
            $ref: '#/definitions/models.Track'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: CreateTrack
      tags:
        - track
    put:
      consumes:
      - application/json
      description: updating track by id
      parameters:
      - description: id of updating song and params of it.
        in: body
        name: Track
        required: true
        schema:
          $ref: '#/definitions/models.Track'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: UpdateTrack
      tags:
        - track
  /net/v1/track/{id}:
    delete:
      consumes:
        - application/json
      description: deleting track by id
      parameters:
        - description: id of track which need to be deleted
          in: path
          name: id
          required: true
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: DeleteTrack
      tags:
        - track
    get:
      consumes:
      - application/json
      description: getting track by id
      parameters:
        - description: id of track which need to be getted
          in: path
          name: id
          required: true
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Track'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetTrack
      tags:
        - track
  /net/v1/track/popular:
    get:
      consumes:
        - application/json
      description: getting top20 popular tracks
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Success'
        "400":
          description: Data is invalid
          schema:
            $ref: '#/definitions/utils.Error'
        "405":
          description: Method is not allowed
          schema:
            $ref: '#/definitions/utils.Error'
      summary: GetPopularTracks
      tags:
        - track
  /net/v1/users/{id}:
    get:
      consumes:
        - application/json
      description: get user by cookie
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.User'
        "400":
          description: invalid id
          schema:
            $ref: '#/definitions/forms.Result'
        "404":
          description: user not found
          schema:
            $ref: '#/definitions/forms.Result'
      summary: GetUser
      tags:
      - user
  /net/v1/users/self:
    get:
      consumes:
        - application/json
      description: get user
      parameters:
        - description: id of user which need to be getted
          in: path
          name: id
          required: true
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/forms.User'
        "401":
          description: invalid csrf
          schema:
            $ref: '#/definitions/forms.Result'
      summary: GetSelfUser
      tags:
      - user
swagger: "2.0"